% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assert_choice}
\alias{assert_choice}
\title{Assert that a value is one of a set of valid choices}
\usage{
assert_choice(x, choices, arg = deparse(substitute(x)))
}
\arguments{
\item{x}{A scalar value to check.}

\item{choices}{Character vector of valid values.}

\item{arg}{Name of the argument, used in the error message.}
}
\value{
\code{x} invisibly if the assertion passes.
}
\description{
Aborts with an informative error listing the valid options if \code{x} is not
contained in \code{choices}. More informative than \code{\link[base:match.arg]{base::match.arg()}} for
pipeline error messages.
}
\examples{
assert_choice("linear", c("linear", "locf", "constant"))
}
