% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assert_length}
\alias{assert_length}
\title{Assert that an object has a specific length}
\usage{
assert_length(x, n, arg = deparse(substitute(x)))
}
\arguments{
\item{x}{The object to check.}

\item{n}{Expected length (single positive integer).}

\item{arg}{Name of the argument, used in the error message.}
}
\value{
\code{x} invisibly if the assertion passes.
}
\description{
Aborts if \code{length(x) != n}. More specific than \code{\link[=assert_scalar]{assert_scalar()}} when an
exact fixed length is required.
}
\examples{
assert_length(c(0, 1), n = 2L)
}
