% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assert_numeric}
\alias{assert_numeric}
\title{Assert that an object is numeric}
\usage{
assert_numeric(x, arg = deparse(substitute(x)), allow_na = TRUE)
}
\arguments{
\item{x}{The object to check.}

\item{arg}{Name of the argument, used in the error message.}

\item{allow_na}{Logical. If \code{FALSE}, aborts when \code{x} contains any \code{NA}.
Default \code{TRUE}.}
}
\value{
\code{x} invisibly if the assertion passes.
}
\description{
Aborts with an informative error if \code{x} is not a numeric vector. Optionally
also aborts if \code{x} contains any \code{NA} values.
}
\examples{
assert_numeric(c(1.2, 3.4, 5.6))
assert_numeric(c(1, 2, NA), allow_na = TRUE)
}
