% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assert_scalar}
\alias{assert_scalar}
\title{Assert that an object has length one}
\usage{
assert_scalar(x, arg = deparse(substitute(x)))
}
\arguments{
\item{x}{The object to check.}

\item{arg}{Name of the argument, used in the error message.}
}
\value{
\code{x} invisibly if the assertion passes.
}
\description{
Aborts if \code{length(x) != 1}. Useful for guarding scalar arguments such as
thresholds and configuration values.
}
\examples{
assert_scalar(42)
assert_scalar("hourly")
}
