% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{floor_to}
\alias{floor_to}
\title{Floor a datetime vector to a given time unit}
\usage{
floor_to(x, unit)
}
\arguments{
\item{x}{A POSIXct vector.}

\item{unit}{A string accepted by \code{\link[lubridate:round_date]{lubridate::floor_date()}}, such as
\code{"15 minutes"}, \code{"1 hour"}, \code{"day"}.}
}
\value{
A POSIXct vector floored to \code{unit}.
}
\description{
A thin wrapper around \code{\link[lubridate:round_date]{lubridate::floor_date()}} using the same interface
as the rest of the reach.utils datetime helpers.
}
\examples{
x <- as_utc("2024-01-15 06:37:22")
floor_to(x, "15 minutes")
floor_to(x, "1 hour")
}
