% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{parse_datetime}
\alias{parse_datetime}
\title{Parse datetimes using common EA formats}
\usage{
parse_datetime(x)
}
\arguments{
\item{x}{A character vector of datetime strings.}
}
\value{
A POSIXct vector in UTC. Values that could not be parsed are \code{NA}.
}
\description{
Attempts to parse datetime strings using a prioritised list of formats
commonly encountered in Environment Agency hydrometric data, falling back
gracefully if a format does not match.
}
\examples{
parse_datetime("2024-01-15T06:00:00Z")
parse_datetime("15/01/2024 06:00")
}
