% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reach.utils-package.R
\docType{package}
\name{reach.utils-package}
\alias{reach.utils}
\alias{reach.utils-package}
\title{reach.utils: Utility Functions for the Flode Ecosystem}
\description{
Provides shared utility functions for the reaches package ecosystem, including datetime helpers, YAML/JSON configuration loading, and structured console and file logging via cli.
}
\seealso{
Useful links:
\itemize{
\item \url{https://github.com/JonPayneEA/reach.utils}
\item Report bugs at \url{https://github.com/JonPayneEA/reach.utils/issues}
}
}
\author{
\strong{Maintainer}: Forecasting and Warning Team \email{forecasting@environment-agency.gov.uk}

}
\keyword{internal}
